+++ src/Form/Type/DateTimeType.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SwagIndustries\Melodiia\Form\Type;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\DateTimeType as OriginalDateTimeType;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class DateTimeType extends AbstractType
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver
+            ->setDefaults([
+                'widget' => 'single_text',
+                'format' => OriginalDateTimeType::HTML5_FORMAT,
+                'input' => 'datetime_immutable',
+            ])
+        ;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getParent()
+    {
+        return OriginalDateTimeType::class;
+    }
+}